Move VC/ESX thumbprint code under broker
