+++ lib/gems/pending/VMwareWebService/esx_thumb_print.rb
