+++ lib/gems/pending/VMwareWebService/vcenter_thumb_print.rb
